refactor: remove as funções do main e coloca cada uma no arquivo que é chamada removendo a necessidade de props nesses arquivos
