+++ app/frontend/src/components/eventos/buttonsUp.jsx
@@ -1,14 +1,73 @@
 import React, { useContext } from 'react'
-import PropTypes from 'prop-types'
 import Context from '../providers/context'
 
 const ButtonsUp = (props) => {
   const {
     text,
+    list,
     btnAdd,
     setText,
-    btnDelete
+    setList,
+    btnDelete,
+    nodeElement,
+    setBtnDelete
   } = useContext(Context)
+
+  const addText = () => {
+    setList([...list,
+      {
+        id: list.length,
+        newText: text,
+        stts: ''
+      }
+    ])
+    setText('')
+  }
+
+  const deleteText = () => {
+    nodeElement.forEach((item) => {
+      if (item.checked) {
+        item.parentNode.remove()
+      }
+    })
+    setBtnDelete(true)
+  }
+
+  const sortABCD = () => {
+    list.sort((a, b) => {
+      if (a.newText > b.newText) return 1
+      if (a.newText < b.newText) return -1
+
+      setList([...list])
+      return 0
+    })
+  }
+
+  // const sortDATA = () => {
+  //   list.sort((a, b) => {
+  //     if (a.id > b.id) return 1
+  //     if (a.id < b.id) return -1
+
+  //     setList([...list])
+  //     return 0
+  //   })
+  // }
+
+  // const sortSTTS = () => {
+  //   list.sort((a, b) => {
+  //     if (a.stts > b.stts) return 1
+  //     if (a.stts < b.stts) return -1
+
+  //     setList([...list])
+  //     return 0
+  //   })
+  // }
+
+  const sortList = (order) => {
+    if (order === 'abcd') sortABCD()
+    // if (order === 'data') sortDATA()
+    // if (order === 'stts') sortSTTS()
+  }
 
   return (
     <section className='btnsUp'>
@@ -22,31 +81,27 @@ const ButtonsUp = (props) => {
         />
         <button
           disabled={ btnAdd }
-          onClick={ props.addText }
+          onClick={ addText }
         >
           Adicionar
         </button>
         <button
           disabled={ btnDelete }
-          onClick={ props.deleteText }
+          onClick={ deleteText }
         >
           Remover
         </button>
       </div>
-      <label htmlFor='select'>Ordenar por:</label>
-      <select id='select' onChange={ (e) => props.sortList(e.target.value) }>
+      <label htmlFor='select'>
+        Ordenar por:
+      </label>
+      <select id='select' onChange={ (e) => sortList(e.target.value) }>
         <option value='abcd'>Alfabética</option>
         <option value='data'>Data de criação</option>
         <option value='stts'>Status</option>
       </select>
     </section>
   )
-}
-
-ButtonsUp.propTypes = {
-  addText: PropTypes.func,
-  sortList: PropTypes.func,
-  deleteText: PropTypes.func
 }
 
 export default ButtonsUp
